Move calculation of integrals to beginning of simulation
We currently only have R R = R Delta, rho rho, rho h1, so the code was
simplified accordingly. If we introduce other RR interactions, the code will
have to be adapted accordingly.

The tests were updated accordingly.

+++ src/decaytype.cc
@@ -142,8 +142,8 @@ double TwoBodyDecaySemistable::get_Lambda() {
   }
 }
 
-// number of tabulation points
-constexpr int num_tab_pts = 200;
+/// Number of tabulation points.
+constexpr size_t num_tab_pts = 200;
 static /*thread_local (see #3075)*/ Integrator integrate;
 
 double TwoBodyDecaySemistable::rho(double mass) const {
